Move LockMaster to its own files
