+++ generate/templates/templates/struct_content.cc
@@ -15,7 +15,7 @@ extern "C" {
 }
 
 #include <iostream>
-#include "../include/nodegit.h"
+#include "../include/lock_master.h"
 #include "../include/functions/copy.h"
 #include "../include/{{ filename }}.h"
 #include "../include/functions/sleep_for_ms.h"
